requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
pyotp==2.9.0
qrcode==7.4.2
email-validator==2.1.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9